requirements.txt
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
DateTime==4.3
defusedxml==0.7.0rc1
Django==2.2.17
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.26.2
zope.interface==5.2.0
mysqlclient==2.0.1
django-mysql==3.8.1
